+++ Sources/Hyperspace/Request/Request.swift
@@ -13,6 +13,7 @@ public struct Request<Response, Error: TransportFailureRepresentable>: Recoverab
     
     // MARK: - Typealias
     public typealias Transformer = (TransportSuccess) -> Result<Response, Error>
+    public typealias RecoveryTransformer = (TransportFailure) -> TransportSuccess?
     public typealias DecodingFailureTransformer = (DecodingFailure) -> Error
     
     // MARK: - Properties
@@ -46,6 +47,9 @@ public struct Request<Response, Error: TransportFailureRepresentable>: Recoverab
     
     /// Attempts to parse the provided `TransportSuccess` into the associated response model type for this request.
     public var successTransformer: Transformer
+
+    /// Attempts to recover from a failure by converting a `TransportFailure` into a `TransportSucces`. The default implementation fails by returning nil.
+    public var recoveryTransformer: RecoveryTransformer = { _ in nil }
     
     // MARK: - Initializer
     
@@ -81,6 +85,13 @@ public struct Request<Response, Error: TransportFailureRepresentable>: Recoverab
         }
     }
 
+    public func map<New>(_ responseTransformer: @escaping (TransportSuccess, Response) -> New) -> Request<New, Error> {
+        return Request<New, Error>(method: method, url: url, headers: headers, body: body, cachePolicy: cachePolicy, timeout: timeout) { transportSuccess in
+            let responseResult = self.transform(success: transportSuccess)
+            return responseResult.map { responseTransformer(transportSuccess, $0) }
+        }
+    }
+
     public func mapError<New>(_ errorTransformer: @escaping (Error) -> New) -> Request<Response, New> {
         return Request<Response, New>(method: method, url: url, headers: headers, body: body, cachePolicy: cachePolicy, timeout: timeout) { transportSuccess in
             let originalResponse = self.transform(success: transportSuccess)
@@ -107,7 +118,7 @@ public struct RequestDefaults {
     public static var defaultCachePolicy: URLRequest.CachePolicy = .useProtocolCachePolicy
     public static var defaultDecoder: JSONDecoder = JSONDecoder()
     public static var defaultMaxRecoveryAttempts: UInt = 1
-    public static var defaultTimeout: TimeInterval = 30
+    public static var defaultTimeout: TimeInterval = 60
 }
 
 // MARK: - Request Default Implementations